Removed some unnecessary casts
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@645 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/cpu/Flags.java
@@ -72,7 +72,7 @@ public class Flags {
     }
 
     public static void lf_var1d(int v) {
-        lflags.var1=(int)v;
+        lflags.var1=v;
     }
 
     public static int lf_var2d() {
